# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LexAI - AI Legal Assistant Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents:
    get:
      operationId: getDocuments
      tags: [documents]
      summary: Get all documents for current user
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{documentId}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /documents/{documentId}/analyze:
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: Analyze a document with AI (summarize, simplify, detect risks, translate)
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeDocumentBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnalysis"

  # LAWYERS
  /lawyers:
    get:
      operationId: getLawyers
      tags: [lawyers]
      summary: Get list of lawyers with optional filters
      parameters:
        - name: specialization
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: minRating
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of lawyers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LawyerCard"
  /lawyers/{lawyerId}:
    get:
      operationId: getLawyer
      tags: [lawyers]
      summary: Get lawyer profile by ID
      parameters:
        - name: lawyerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lawyer profile detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LawyerDetail"
  /lawyers/{lawyerId}/reviews:
    get:
      operationId: getLawyerReviews
      tags: [lawyers]
      summary: Get reviews for a lawyer
      parameters:
        - name: lawyerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: addLawyerReview
      tags: [lawyers]
      summary: Add a review for a lawyer
      parameters:
        - name: lawyerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddReviewBody"
      responses:
        "201":
          description: Review added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /lawyers/{lawyerId}/availability:
    get:
      operationId: getLawyerAvailability
      tags: [bookings]
      summary: Get available time slots for a lawyer
      parameters:
        - name: lawyerId
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSlot"

  # MEMBERSHIP
...
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        extractedText:
          type: string
        summary:
          type: string
        uploadedAt:
          type: string
      required: [id, userId, name, fileType, uploadedAt]
    UploadDocumentBody:
      type: object
      properties:
        name:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        content:
          type: string
      required: [name, fileType]
    AnalyzeDocumentBody:
      type: object
      properties:
        action:
          type: string
          enum: [summarize, simplify, detect_risks, translate, ask_question]
        question:
          type: string
        targetLanguage:
          type: string
      required: [action]
    DocumentAnalysis:
      type: object
      properties:
        documentId:
          type: integer
        action:
          type: string
        result:
          type: string
        disclaimer:
          type: string
      required: [documentId, action, result]

    # LAWYER SCHEMAS
    LawyerCard:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        specialization:
          type: string
        languages:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        hourlyRate:
          type: number
        verificationStatus:
          type: string
      required: [id, userId, firstName, lastName, specialization, languages, yearsOfExperience, rating, reviewCount]
    LawyerDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        specialization:
          type: string
        languages:
          type: array
          items:
            type: string
        yearsOfExperience:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        hourlyRate:
          type: number
        bio:
          type: string
        barNumber:
          type: string
        verificationStatus:
          type: string
        availability:
          type: array
          items:
            type: string
      required: [id, userId, firstName, lastName, specialization, languages, yearsOfExperience, rating, reviewCount]
    Review:
      type: object
      properties:
        id:
          type: integer
        lawyerId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
      required: [id, lawyerId, userId, rating, createdAt]
    AddReviewBody:
      type: object
      properties:
        rating:
          type: number
        comment:
          type: string
      required: [rating]
...
    TimeSlot:
      type: object
      properties:
        startTime:
          type: string
        endTime:
          type: string
        available:
          type: boolean
      required: [startTime, endTime, available]

    # MEMBERSHIP SCHEMAS
...